Refactory job deployable factory
